Fixing bugs with aperture

+++ src/drivers/portal.driver.ts
@@ -9,10 +9,12 @@ import { closePortalInstance, openPortalInstance } from './router'
 export const createPortal = (request: PortalRequest) => new Promise<Portal>(async (resolve, reject) => {
 		try {
 			const portal = await new Portal().create(request)
-			const mountpoint = await new Mountpoint().create(portal)
-
+			if(process.env.JANUS_ENABLE === 'true') {
+				const mountpoint = await new Mountpoint().create(portal)
+				createJanusStreamingMountpoint(mountpoint)
+			}
+			
 			openPortalInstance(portal)
-			createJanusStreamingMountpoint(mountpoint)
 
 			resolve(portal)
 		} catch (error) {
@@ -23,12 +25,15 @@ export const createPortal = (request: PortalRequest) => new Promise<Portal>(asyn
 export const closePortal = (portalId: string) => new Promise(async (resolve, reject) => {
 	try {
 		const portal = await new Portal().load(portalId)
-		const mountpoint = await new Mountpoint().load('Portal', portalId)
 		await portal.destroy()
-		await mountpoint.destroy()
+		
+		if(process.env.JANUS_ENABLE === 'true') {
+			const mountpoint = await new Mountpoint().load('Portal', portalId)
+			destroyJanusStramingMountpoint(mountpoint)
+			await mountpoint.destroy()
+		}
 
 		closePortalInstance(portal)
-		destroyJanusStramingMountpoint(mountpoint)
 
 		if (portal.status === 'open')
 			checkNextQueueItem()
